+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Tweak stat: participants by date can be found in the past
+
 ### 1.8.7: November 2nd, 2021
 * Booking Form: validate postal code field (only digits and 5 characters)
 * Booking Form: validate city field (cannot have digits)